pyproject.toml
[tool.poetry]
name = "fastapi-todo"
version = "0.1.0"
description = ""
authors = ["KJD <125159447+Thrylos77@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.118.2"
uvicorn = "^0.37.0"
sqlalchemy = "^2.0.43"
alembic = "^1.16.5"
psycopg2-binary = "^2.9.10"
slowapi = "^0.1.9"
python-dotenv = "^1.1.1"
pyjwt = "^2.10.1"
passlib = "^1.7.4"
bcrypt = "^5.0.0"
python-multipart = "^0.0.20"
email-validator = "^2.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-asyncio = "^1.2.0"
httpx = "^0.28.1"
black = "^25.9.0"
ruff = "^0.14.0"
mypy = "^1.18.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
